fix: Dialog stories
